# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lamontville Golden Arrows FC API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface NewsArticle {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  /** @nullable */
  content?: string | null;
  category: string;
  imageUrl: string;
  featured?: boolean;
  author?: string;
  tags?: string[];
  publishedAt: string;
}

export interface NewsInput {
  title: string;
  slug: string;
  excerpt: string;
  content?: string;
  category: string;
  imageUrl: string;
  featured?: boolean;
  author?: string;
  tags?: string[];
}

export interface Player {
  id: number;
  name: string;
  position: string;
  number: number;
  nationality: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  appearances?: number;
  goals?: number;
  assists?: number;
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  facebook?: string | null;
  /** @nullable */
  twitter?: string | null;
}

export interface PlayerInput {
  name: string;
  position: string;
  number: number;
  nationality: string;
  age?: number;
  photoUrl?: string;
  bio?: string;
  appearances?: number;
  goals?: number;
  assists?: number;
  instagram?: string;
  facebook?: string;
  twitter?: string;
}

export interface StaffMember {
  id: number;
  name: string;
  role: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  nationality?: string | null;
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  facebook?: string | null;
  /** @nullable */
  twitter?: string | null;
}

export interface StaffInput {
  name: string;
  role: string;
  photoUrl?: string;
  bio?: string;
  nationality?: string;
  instagram?: string;
  facebook?: string;
  twitter?: string;
}

export interface Fixture {
  id: number;
  date: string;
  /** @nullable */
  time?: string | null;
  homeTeam: string;
  awayTeam: string;
  competition: string;
  venue: string;
  /** @nullable */
  ticketUrl?: string | null;
}

export interface MatchResult {
  id: number;
  date: string;
  homeTeam: string;
  awayTeam: string;
  homeScore: number;
  awayScore: number;
  competition: string;
  /** @nullable */
  venue?: string | null;
  scorers?: string[];
  /** @nullable */
  matchReport?: string | null;
  /** @nullable */
  highlightUrl?: string | null;
}

export interface LeagueTableEntry {
  season?: number;
  position: number;
  team: string;
  /** @nullable */
  logoUrl?: string | null;
  played: number;
  won: number;
  drawn: number;
  lost: number;
  goalsFor: number;
  goalsAgainst: number;
  goalDifference: number;
  points: number;
  isGoldenArrows?: boolean;
}

export interface LeagueTableResponse {
  season: number;
  entries: LeagueTableEntry[];
}

export type GalleryItemType = typeof GalleryItemType[keyof typeof GalleryItemType];


export const GalleryItemType = {
  photo: 'photo',
  video: 'video',
} as const;

export interface GalleryItem {
  id: number;
  title: string;
  type: GalleryItemType;
  url: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  category: string;
  /** @nullable */
  caption?: string | null;
  publishedAt?: string;
}

export type GalleryInputType = typeof GalleryInputType[keyof typeof GalleryInputType];


export const GalleryInputType = {
  photo: 'photo',
  video: 'video',
} as const;

export interface GalleryInput {
  title: string;
  type: GalleryInputType;
  url: string;
  thumbnailUrl?: string;
  category: string;
  caption?: string;
}

export interface Sponsor {
  id: number;
  name: string;
  logoUrl: string;
  /** @nullable */
  websiteUrl?: string | null;
  tier: string;
}

export interface Slide {
  id: number;
  title: string;
  imageUrl: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  link?: string | null;
  /** @nullable */
  linkLabel?: string | null;
  sortOrder: number;
  active: boolean;
  createdAt: string;
}

export interface SlideInput {
  title: string;
  imageUrl: string;
  subtitle?: string;
  link?: string;
  linkLabel?: string;
  sortOrder?: number;
  active?: boolean;
}

export interface SlideUpdate {
  title?: string;
  imageUrl?: string;
  subtitle?: string;
  link?: string;
  linkLabel?: string;
  sortOrder?: number;
  active?: boolean;
}

export interface MatchResultInput {
  date: string;
  homeTeam: string;
  awayTeam: string;
  homeScore: number;
  awayScore: number;
  competition: string;
  venue?: string;
  scorers?: string[];
  matchReport?: string;
  highlightUrl?: string;
}

export type EnquiryStatus = typeof EnquiryStatus[keyof typeof EnquiryStatus];


export const EnquiryStatus = {
  unread: 'unread',
  read: 'read',
  resolved: 'resolved',
} as const;

export interface Enquiry {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  subject: string;
  message: string;
  status: EnquiryStatus;
  createdAt: string;
}

export interface EnquiryInput {
  firstName: string;
  lastName: string;
  email: string;
  subject: string;
  message: string;
}

export type EnquiryUpdateStatus = typeof EnquiryUpdateStatus[keyof typeof EnquiryUpdateStatus];


export const EnquiryUpdateStatus = {
  unread: 'unread',
  read: 'read',
  resolved: 'resolved',
} as const;

export interface EnquiryUpdate {
  status?: EnquiryUpdateStatus;
}

export interface SyncResult {
  synced: number;
  total: number;
  season: number;
  note?: string;
}

export interface StatsSummary {
  leaguePosition: number;
  played: number;
  won: number;
  drawn: number;
  lost: number;
  goalsScored: number;
  goalsConceded: number;
  points: number;
  topScorer: string;
  topScorerGoals?: number;
  cleanSheets: number;
  winRate?: number;
}

export interface Ad {
  id: number;
  slot: string;
  imageUrl?: string | null;
  linkUrl?: string | null;
  altText?: string | null;
  updatedAt: string;
}

export interface AdUpdate {
  imageUrl?: string | null;
  linkUrl?: string | null;
  altText?: string | null;
}

export type ListEnquiriesParams = {
status?: string;
limit?: number;
offset?: number;
};

export type ListNewsParams = {
category?: string;
featured?: boolean;
limit?: number;
offset?: number;
};

export type ListPlayersParams = {
position?: string;
};

export type ListFixturesParams = {
competition?: string;
limit?: number;
};

export type ListResultsParams = {
limit?: number;
};

export type ListGalleryParams = {
category?: string;
type?: ListGalleryType;
limit?: number;
};

export type ListGalleryType = typeof ListGalleryType[keyof typeof ListGalleryType];


export const ListGalleryType = {
  photo: 'photo',
  video: 'video',
} as const;