# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lamontville Golden Arrows FC API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ads/{slot}:
    get:
      operationId: getAd
      tags: [ads]
      summary: Get a single ad slot
      parameters:
        - in: path
          name: slot
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Ad slot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
        "404":
          description: Not found
    put:
      operationId: updateAd
      tags: [ads]
      summary: Update an ad slot (admin)
      parameters:
        - in: path
          name: slot
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdUpdate"
      responses:
        "200":
          description: Updated ad slot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
...
  /enquiries:
    post:
      operationId: createEnquiry
      tags: [enquiries]
      summary: Submit a contact enquiry (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnquiryInput"
      responses:
        "201":
          description: Enquiry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"
    get:
      operationId: listEnquiries
      tags: [enquiries]
      summary: List all enquiries (admin)
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: offset
          schema: { type: integer }
      responses:
        "200":
          description: List of enquiries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enquiry"
  /enquiries/{id}:
    patch:
      operationId: updateEnquiry
      tags: [enquiries]
      summary: Update enquiry status (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnquiryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"
        "404":
          description: Not found
    delete:
      operationId: deleteEnquiry
      tags: [enquiries]
      summary: Delete an enquiry (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── NEWS ──────────────────────────────────────────────────────────────────
  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List news articles
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: featured
          schema: { type: boolean }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: offset
          schema: { type: integer }
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsArticle"
    post:
      operationId: createNews
      tags: [news]
      summary: Create news article (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
  /news/{id}:
    get:
      operationId: getNewsArticle
      tags: [news]
      summary: Get single news article
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
        "404":
          description: Not found
    put:
      operationId: updateNews
      tags: [news]
      summary: Update news article (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
    delete:
      operationId: deleteNews
      tags: [news]
      summary: Delete news article (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── PLAYERS ───────────────────────────────────────────────────────────────
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List squad players
      parameters:
        - in: query
          name: position
          schema: { type: string }
      responses:
        "200":
          description: Players list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
    post:
      operationId: createPlayer
      tags: [players]
      summary: Create player (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players/{id}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get player profile
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: Not found
    put:
      operationId: updatePlayer
      tags: [players]
      summary: Update player (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
    delete:
      operationId: deletePlayer
      tags: [players]
      summary: Delete player (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── STAFF ─────────────────────────────────────────────────────────────────
...
  /gallery:
    get:
      operationId: listGallery
      tags: [gallery]
      summary: List gallery items
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string, enum: [photo, video] }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: Gallery items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryItem"
    post:
      operationId: createGalleryItem
      tags: [gallery]
      summary: Upload gallery item (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
  /gallery/{id}:
    delete:
      operationId: deleteGalleryItem
      tags: [gallery]
      summary: Delete gallery item (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── SPONSORS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    NewsArticle:
      type: object
      required: [id, title, slug, category, publishedAt, imageUrl, excerpt]
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string }
        content: { type: ["string", "null"] }
        category: { type: string }
        imageUrl: { type: string }
        featured: { type: boolean }
        author: { type: string }
        tags: { type: array, items: { type: string } }
        publishedAt: { type: string }
    NewsInput:
      type: object
      required: [title, slug, category, imageUrl, excerpt]
      properties:
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string }
        content: { type: string }
        category: { type: string }
        imageUrl: { type: string }
        featured: { type: boolean }
        author: { type: string }
        tags: { type: array, items: { type: string } }

    # ── PLAYERS ───────────────────────────────────────────────────────────
    Player:
      type: object
      required: [id, name, position, number, nationality]
      properties:
        id: { type: integer }
        name: { type: string }
        position: { type: string }
        number: { type: integer }
        nationality: { type: string }
        age: { type: ["integer", "null"] }
        photoUrl: { type: ["string", "null"] }
        bio: { type: ["string", "null"] }
        appearances: { type: integer }
        goals: { type: integer }
        assists: { type: integer }
        instagram: { type: ["string", "null"] }
        facebook: { type: ["string", "null"] }
        twitter: { type: ["string", "null"] }
    PlayerInput:
      type: object
      required: [name, position, number, nationality]
      properties:
        name: { type: string }
        position: { type: string }
        number: { type: integer }
        nationality: { type: string }
        age: { type: integer }
        photoUrl: { type: string }
        bio: { type: string }
        appearances: { type: integer }
        goals: { type: integer }
        assists: { type: integer }
        instagram: { type: string }
        facebook: { type: string }
        twitter: { type: string }

    # ── STAFF ─────────────────────────────────────────────────────────────
...
    GalleryItem:
      type: object
      required: [id, title, type, url, category]
      properties:
        id: { type: integer }
        title: { type: string }
        type: { type: string, enum: [photo, video] }
        url: { type: string }
        thumbnailUrl: { type: ["string", "null"] }
        category: { type: string }
        caption: { type: ["string", "null"] }
        publishedAt: { type: string }
    GalleryInput:
      type: object
      required: [title, type, url, category]
      properties:
        title: { type: string }
        type: { type: string, enum: [photo, video] }
        url: { type: string }
        thumbnailUrl: { type: string }
        category: { type: string }
        caption: { type: string }

    # ── SPONSORS ──────────────────────────────────────────────────────────
...
    Enquiry:
      type: object
      required: [id, firstName, lastName, email, subject, message, status, createdAt]
      properties:
        id: { type: integer }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        subject: { type: string }
        message: { type: string }
        status: { type: string, enum: [unread, read, resolved] }
        createdAt: { type: string }
    EnquiryInput:
      type: object
      required: [firstName, lastName, email, subject, message]
      properties:
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        subject: { type: string }
        message: { type: string }
    EnquiryUpdate:
      type: object
      properties:
        status: { type: string, enum: [unread, read, resolved] }

    # ── SYNC ──────────────────────────────────────────────────────────────
...
    Ad:
      type: object
      required: [id, slot, updatedAt]
      properties:
        id: { type: integer }
        slot: { type: string }
        imageUrl: { type: string, nullable: true }
        linkUrl: { type: string, nullable: true }
        altText: { type: string, nullable: true }
        updatedAt: { type: string }
    AdUpdate:
      type: object
      properties:
        imageUrl: { type: string, nullable: true }
        linkUrl: { type: string, nullable: true }
        altText: { type: string, nullable: true }